requirements.txt
opencv-python
numpy
pillow
ultralytics
torch
